+++ src/test/groovy/de/thomaskrille/dropwizard_template_config/AdvancedInterpolationSpec.groovy
@@ -1,5 +1,6 @@
 package de.thomaskrille.dropwizard_template_config
 
+import com.google.common.base.Charsets
 import org.apache.commons.io.IOUtils
 import spock.lang.Specification
 
@@ -12,7 +13,8 @@ class AdvancedInterpolationSpec extends Specification {
     def TemplateConfigurationSourceProvider templateConfigurationSourceProvider =
             new TemplateConfigurationSourceProvider(new TestConfigSourceProvider(),
                     environmentProvider,
-                    new DefaultSystemPropertiesProvider())
+                    new DefaultSystemPropertiesProvider(),
+                    Charsets.UTF_8)
 
     def 'replacing an environment variable inline works'() {
         given: